Move creation of validators and organizations to mock and pass them into the tests
