+++ pallets/stellar-relay/src/mock.rs
@@ -1,15 +1,24 @@
 use frame_support::{
+	pallet_prelude::GenesisBuild,
 	parameter_types,
 	traits::{ConstU16, ConstU64},
+	BoundedVec,
 };
 use frame_system as system;
+use rand::Rng;
 use sp_core::H256;
 use sp_runtime::{
 	testing::Header,
 	traits::{BlakeTwo256, IdentityLookup},
 };
+use substrate_stellar_sdk::SecretKey;
 
 use crate as pallet_spacewalk_relay;
+use crate::{
+	traits::{FieldLength, Organization, Validator},
+	types::{OrganizationOf, ValidatorOf},
+	Error,
+};
 
 type UncheckedExtrinsic = frame_system::mocking::MockUncheckedExtrinsic<Test>;
 type Block = frame_system::mocking::MockBlock<Test>;
@@ -68,7 +77,130 @@ impl pallet_spacewalk_relay::Config for Test {
 	type WeightInfo = ();
 }
 
-// Build genesis storage according to the mock runtime.
-pub fn new_test_ext() -> sp_io::TestExternalities {
-	system::GenesisConfig::default().build_storage::<Test>().unwrap().into()
+fn create_bounded_vec<T: Clone>(input: &[T]) -> Result<BoundedVec<T, FieldLength>, Error<Test>> {
+	let bounded_vec = BoundedVec::try_from(input.to_vec())
+		.map_err(|_| Error::<Test>::BoundedVecCreationFailed)?;
+	Ok(bounded_vec)
+}
+
+pub fn create_dummy_validator(
+	name: &str,
+	organization: &OrganizationOf<Test>,
+) -> (ValidatorOf<Test>, SecretKey) {
+	let rand = &mut rand::thread_rng();
+	let validator_secret = SecretKey::from_binary(rand.gen());
+
+	let validator = Validator {
+		name: create_bounded_vec(name.as_bytes()).unwrap(),
+		public_key: create_bounded_vec(validator_secret.get_public().to_encoding().as_slice())
+			.unwrap(),
+		organization_id: organization.id.clone(),
+	};
+
+	(validator, validator_secret)
+}
+
+fn create_dummy_validators() -> (Vec<OrganizationOf<Test>>, Vec<ValidatorOf<Test>>, Vec<SecretKey>)
+{
+	let mut organizations: Vec<OrganizationOf<Test>> = vec![];
+	let mut validators: Vec<ValidatorOf<Test>> = vec![];
+	// These secret keys are required to be in the same order as the validators in this test
+	// They are later used to sign the dummy scp messages
+	let mut validator_secret_keys: Vec<SecretKey> = vec![];
+
+	let organization_sdf =
+		Organization { name: create_bounded_vec("sdf".as_bytes()).unwrap(), id: 0 };
+	organizations.push(organization_sdf.clone());
+
+	let (validator, validator_secret) = create_dummy_validator("$sdf1", &organization_sdf);
+	validators.push(validator);
+	validator_secret_keys.push(validator_secret);
+	let (validator, validator_secret) = create_dummy_validator("$sdf2", &organization_sdf);
+	validators.push(validator);
+	validator_secret_keys.push(validator_secret);
+	let (validator, validator_secret) = create_dummy_validator("$sdf3", &organization_sdf);
+	validators.push(validator);
+	validator_secret_keys.push(validator_secret);
+
+	let organization_keybase =
+		Organization { name: create_bounded_vec("keybase".as_bytes()).unwrap(), id: 1 };
+	organizations.push(organization_keybase.clone());
+
+	let (validator, validator_secret) = create_dummy_validator("$keybase1", &organization_keybase);
+	validators.push(validator);
+	validator_secret_keys.push(validator_secret);
+	let (validator, validator_secret) = create_dummy_validator("$keybase2", &organization_keybase);
+	validators.push(validator);
+	validator_secret_keys.push(validator_secret);
+	let (validator, validator_secret) = create_dummy_validator("$keybase3", &organization_keybase);
+	validators.push(validator);
+	validator_secret_keys.push(validator_secret);
+
+	let organization_satoshipay =
+		Organization { name: create_bounded_vec("satoshipay".as_bytes()).unwrap(), id: 2 };
+	organizations.push(organization_satoshipay.clone());
+
+	let (validator, validator_secret) =
+		create_dummy_validator("$satoshipay-de", &organization_satoshipay);
+	validators.push(validator);
+	validator_secret_keys.push(validator_secret);
+	let (validator, validator_secret) =
+		create_dummy_validator("$satoshipay-us", &organization_satoshipay);
+	validators.push(validator);
+	validator_secret_keys.push(validator_secret);
+	let (validator, validator_secret) =
+		create_dummy_validator("$satoshipay-sg", &organization_satoshipay);
+	validators.push(validator);
+	validator_secret_keys.push(validator_secret);
+
+	let organization_lobstr =
+		Organization { name: create_bounded_vec("lobstr".as_bytes()).unwrap(), id: 3 };
+	organizations.push(organization_lobstr.clone());
+
+	let (validator, validator_secret) = create_dummy_validator("$lobstr1", &organization_lobstr);
+	validators.push(validator);
+	validator_secret_keys.push(validator_secret);
+	let (validator, validator_secret) = create_dummy_validator("$lobstr2", &organization_lobstr);
+	validators.push(validator);
+	validator_secret_keys.push(validator_secret);
+	let (validator, validator_secret) = create_dummy_validator("$lobstr3", &organization_lobstr);
+	validators.push(validator);
+	validator_secret_keys.push(validator_secret);
+	let (validator, validator_secret) = create_dummy_validator("$lobstr4", &organization_lobstr);
+	validators.push(validator);
+	validator_secret_keys.push(validator_secret);
+
+	(organizations, validators, validator_secret_keys)
+}
+
+pub struct ExtBuilder;
+
+impl ExtBuilder {
+	pub fn build(
+		organizations: Vec<OrganizationOf<Test>>,
+		validators: Vec<ValidatorOf<Test>>,
+	) -> sp_io::TestExternalities {
+		let mut storage = frame_system::GenesisConfig::default().build_storage::<Test>().unwrap();
+
+		crate::GenesisConfig::<Test> {
+			validators,
+			organizations,
+			is_public_network: true,
+			phantom: Default::default(),
+		}
+		.assimilate_storage(&mut storage)
+		.unwrap();
+
+		storage.into()
+	}
+}
+
+pub fn run_test<T>(test: T)
+where
+	T: FnOnce(Vec<OrganizationOf<Test>>, Vec<ValidatorOf<Test>>, Vec<SecretKey>),
+{
+	let (organizations, validators, validator_secret_keys) = create_dummy_validators();
+	ExtBuilder::build(organizations.clone(), validators.clone()).execute_with(|| {
+		test(organizations, validators, validator_secret_keys);
+	});
 }
